fix a few conversion warnings
- fix conversion from large type tp small

(cherry picked from commit 01ebc092d274d5244d2375d049f3baf91750f2a6)

+++ xmrstak/misc/console.hpp
@@ -50,4 +50,4 @@ private:
     FILE* logfile;
 };
 
-void win_exit(size_t code = 1);
+void win_exit(int code = 1);
